+++ src/OpenTelemetry.Api/Context/Propagation/ITextFormat.cs
@@ -39,7 +39,7 @@ namespace OpenTelemetry.Context.Propagation
         /// <param name="context">The default context to transmit over the wire.</param>
         /// <param name="carrier">Object to set context on. Instance of this object will be passed to setter.</param>
         /// <param name="setter">Action that will set name and value pair on the object.</param>
-        void Inject<T>(TextFormatContext context, T carrier, Action<T, string, string> setter);
+        void Inject<T>(PropagationContext context, T carrier, Action<T, string, string> setter);
 
         /// <summary>
         /// Extracts activity context from textual representation.
@@ -49,6 +49,6 @@ namespace OpenTelemetry.Context.Propagation
         /// <param name="carrier">Object to extract context from. Instance of this object will be passed to the getter.</param>
         /// <param name="getter">Function that will return string value of a key with the specified name.</param>
         /// <returns>Context from it's text representation.</returns>
-        TextFormatContext Extract<T>(TextFormatContext context, T carrier, Func<T, string, IEnumerable<string>> getter);
+        PropagationContext Extract<T>(PropagationContext context, T carrier, Func<T, string, IEnumerable<string>> getter);
     }
 }