+++ test/OpenTelemetry.Instrumentation.Http.Tests/HttpWebRequestTests.Basic.netfx.cs
@@ -97,8 +97,8 @@ namespace OpenTelemetry.Instrumentation.Http.Tests
         public async Task HttpWebRequestInstrumentationInjectsHeadersAsync_CustomFormat()
         {
             var textFormat = new Mock<ITextFormat>();
-            textFormat.Setup(m => m.Inject(It.IsAny<TextFormatContext>(), It.IsAny<HttpWebRequest>(), It.IsAny<Action<HttpWebRequest, string, string>>()))
-                .Callback<TextFormatContext, HttpWebRequest, Action<HttpWebRequest, string, string>>((context, message, action) =>
+            textFormat.Setup(m => m.Inject(It.IsAny<PropagationContext>(), It.IsAny<HttpWebRequest>(), It.IsAny<Action<HttpWebRequest, string, string>>()))
+                .Callback<PropagationContext, HttpWebRequest, Action<HttpWebRequest, string, string>>((context, message, action) =>
                 {
                     action(message, "custom_traceparent", $"00/{context.ActivityContext.TraceId}/{context.ActivityContext.SpanId}/01");
                     action(message, "custom_tracestate", Activity.Current.TraceStateString);